implied: fall out from fake_cur_stree changes
This basically only affect check_leaks.c.  That check wants to set the true
false states but then the implication code overwrites it with something
less specific.  This wasn't an issue before when we used to set the cur
stree whenever we set the true path, but now it is.

Fixes: 6e9820f295e2 ('states: __set_sm_cur_stree() should not change the fake_cur_stree()')
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>