implied: unbreak assign_ranged_functions
This is more fallout from the fake_cur_stree changes.  This change is
not intuitive.  We are setting the cur_stree and we rely on
smatch_implied.c to set the true and false states.

Originally we were setting the true and false states ourselves, with wrong
data that was then overwritten by the implied states, but now we don't
overwrite the true and false states if they have been set already.  So this
patch means we don't set them.

It seems complicated.  Heck, it *is* complicated.  But not setting the
true and false states to bad values is the correct thing so this patch is
correct on it's own regardless of other changes.

Fixes: a54ac829a66d ('implied: fall out from fake_cur_stree changes')
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>