implications: use stored_conditions the default fall back
The issue here is say you have a condition like:

	if ((foo & 0xff) == 0x8) {

We were not handling this in smatch_stored_conditions.c because it had a
comparison and we expected that to be handled in smatch_extra.c.  Meanwhile
smatch_extra.c does not record the value of (foo & 0xff) so it has nothing.
Solve this by trying to use smatch_extra first and then stored_conditions
for anything possible.

Using smatch_extra.c is generally better because it knows that "x = 1"
means "x == 1" afterward but smatch_stored_conditions.c only knows about
conditions so it has less information.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>