states: only set the fake_stree in set_state()
I'm trying to untangle how the fake stree is used so that it's a bit
more explicit what's happening.

The complicated bit of this patch is:
-	if (fake_cur_stree_stack)
+	if (fake_cur_stree_stack && owner != -1)

The issue there is that this is called from Smatch implied like this:

        __push_fake_cur_stree();
        __unnullify_path();
        set_extra_nomod(name, sym, alloc_estate_rl(rl));
        extra_states = __pop_fake_cur_stree();

If we don't allow -1 to go through then we can unnullify the path.

Hm...  That seems more complicated than necessary.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>