db: fix a bug in splitting return states
The problem here is that not all the possible states were stored in
->possible so we sometimes missed returns.  For example, in sm_implied15.c
there are two paths which return 0 but Smatch was only seeing one.

One thing to consider for future work is to use smatch_implied.c instead
of implementing implications a different way.  Smatch_implied.c doesn't
use the ->possible list so there are potential memory savings here.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>