flow, conditions: handle label statements correctly
Label statements are basically if you have only one statement after a
label.  Like this:

    out:
	return ret;
    }


If you have more than one statement then there is no label statement.  If
it was the end of a function like the above example, then that's fine it
was already handled correctly.  The problem comes when you have something
like:

	x = ({
		blah();
		blah();
__end:
		val;
	      });

We weren't seeing the last statement and it cause subtle breakage.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>