type: fix "strange non-value function or array" warnings
If you did `kchecker drivers/media/pci/ivtv/ivtv-mailbox.c` then it gives
some warnings like:

include/linux/jiffies.h:359:41: error: strange non-value function or array

It turns out that these are from get_type() when we process
__builtin_constant_p().  I don't know the details of this stuff.  I'm
fixing this by handling it as a special case.  That might actually be the
correct thing to do.  Changing this seems to take care of a bunch of
warnings.

Reported-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>