implied: introduce assume(condition)
What this does is it looks up the implications of a condition and then
sets up a fake cur_stree as if those conditions are true.  If it returns
zero that means it didn't find any implications so it didn't do anything.

Call end_assume() to free the fake cur_stree.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>