db: split positive from negative returns
This is sort of a fallback solution if we have too many states to split it
up into individual return states.  There are quite a few functions which
return negative error code, zero or one and zero and one mean different
things.  So what we are doing here is splitting it up three ways, positive,
negative and zero.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>