implied: remove buggy code for working around the old fake_stree methods
I recently re-worked how set_true_false and set_state worked so this code
is no longer needed.  It had a problem as well because it was calling
set_state_stree() instead of set_sm_state_stree().  So if you had a
&merged state it copied that and you lost the ->possible list.  This
meant that some warnings were suppressed by mistake.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>