flow: fix inside_loop() with inline functions
The inside_loop() function returns true if we have a backwards goto.  It's
used by smatch_impossible because we don't mark things as impossible if
we are inside a loop.  (Smatch handles loops badly).

Anyway, it there was a bug where if we parsed an inline statement with a
backwards goto then it marked the caller as being inside a loop as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>