db: fix bug in print_struct_members
I added a test to bail early if we are not dealing with struct members.

But the other test is a bugfix.  Say you have a state "foobar" and a
param "foo" then it would print that "$bar" was set.  We only want to deal
with struct members where "foo->bar" becomes "$-bar".  This bug is harmless
since the bogus output is ignored but it looks worrisome and is a waste.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>