function_hooks: handle returs that don't set a state
Say you have a "return 0;" with no implications then and another
"return 0;" where it implies that the parameter is 42.  We were ignoring
the first return because no states were being set and only the second
return was used.

Fix this by always setting the path to non-null.

So it goes:
	save original states

	delete all states.
	unnulify the path.
	restore original states
	push a fake stree
	then if the return value is one we are using for the true side
	    set the path to non null which sets something in the fake
	    cur stree and forces a merge.  The merge_fake_cur stree will
	    "match the states" which means that we pull everything in
	    from the cur stree.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>