flow: doen't reset loop_count for inline functions
Say you have an inline function:

void frob(struct foo *foo)
{
	foo->bar++;
}

You want the ++ to be handled differently if you inside an loop or not.

This is probably a deeper issue than I am making it because we have the
same situation for non-inline statements...

Anyway, it fixes the specific bug I'm working on now, and it sort of makes
sense to consider the outer loop.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>