hooks: rename CALL_HOOK_AFTER to CALL_HOOK_AFTER_DB and change it a bit
This is a bugfix.  Back in the day, when there was a function call used in
a condition we used to query the database twice and so
smatch_untracked_param.c worked but now we only do it once.  The problem is
that we call FUNCTION_CALL_HOOK_AFTER before we query the DB but that needs
to be after.

There is the only place that uses CALL_HOOK_AFTER and it really wants to
know what the states are after we'd queried the database.  So I have
renamed it and changed where it is called.

Fixes: 0384c14a42b8 ('function_hooks: don't process condition functions twice')
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>