imaginary_absolute: it's like smatch_extra.c for impossible states
Say you have code like:

	ret = 0;
	if (ret < 0)
		return foo;

This is not far fetched because there are many functions which never fail
but we still check for negative returns.  In the original code Smatch was
recording that we would return s32min-s32max because foo is impossible and
we don't know the value of foo.  Now that we have smatch_imaginary_absolute.c
it says that the return value is negative.

The imaginary_absolute code is very simple.  If you merge of modify
anything it just goes to the empty state.  Which is fine for this purpose.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>