flow: fix bug i just introduced with known switch statements
In the original code if you had an int literal switch statement like
switch (1) then it had a problem.  I'm not certain why.  Before I made this
introduced this bug, it used to work because we failed when expr_to_var_sym
on the expression didn't work.  I just restored the original behavior a
different way.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>