flow: merge multiple case statements in a row together
There is a kernel file where we have like 150 case statements in a row.
I forget which it is.  It was taking a very long time to process so I
thought I would do this to fix it.  Then after I finally wrote the code,
I couldn't find the file and apparently it wasn't in the top ten long
running files anymore.  But this should speed it up anyway.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>