implied: create fake implication history
This patch is not perfect and introduces a couple bugs and exposes a couple
others.  But it's pretty exciting for me because I have been trying to
figure this out for years.

Say we have a function that returns a bunch of negative error codes and
zero on success.  Imagine that smatch groups all the error codes together.
Then we have code like:

    err = frob();
    if (err == -EAGAIN)
        err = -EIO;
    if (err)
        return err;

We almost handle this case because of commit 209b75b277f8 ('implied:
overwrite smatch_extra using implied states if possible').  But because we
said that all the negative returns were grouped together it means that we
set the *mixed and cannot preserve the implications.

Now we can.  What we do is we create fake slist where err is negative but
not -EAGAIN and a fake slist where it is -EAGAIN.  Then we merge them.  We
change the sm state from being a leaf sm to being a merged sm state.

There are a couple complications here.  One is that in smatch_equiv.c we
set up "borrowed" implications.  Meaning that we set the ->left and ->right
sm so that they're pointing to a different variable.  If we run into this
situation then fall back to previous behavior.

The other issue is that sometimes the implied state is less accurate than
the Smatch extra state.  If the implied state is implied because of
smatch_conditions, smatch_compare or the param_limit then fall back to the
previous behavior.

There is a new type introduced somewhere.  Also rl_invert() seems buggy so
rl_intersection() is also buggy.  I will work on these issues in the next
few days, but I wanted to commit what I have so far.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>