implied: pass sm_states instead of pools
This doesn't change anything.  Really we want the sm_state instead of the
pool.  The sm_state has the pool so it's not complicated to preserve the
existing behavior.

The pool list isn't sorted any more so I had to remove the early return in
pool_in_pools().  I don't think we can get duplicates in that list if we just
use add_ptr_list().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>