implied: fix an old implication bug
The highest_stree_id() check was supposed to be an optimization to speed
up the process.  I guess it's supposed to say that if we have two
variables like this:

	foo = frob();

	...

	bar = frob();


Then we can not determine the value of bar from looking at foo because
bar was set later.  There are two problems with this.  First I'm not sure
that's true?  Anyway, the stree_id is incremented when we merge sm_states
so it is definitely wrong as-is.  I found this out looking at some missed
implications in the kernel that led to false positives.

However when I deleted it, then it triggered some failures in validation/
so I needed to replace it with sm_in_keep_leafs().  Which means that if
there is a path from the sm we are looking at to a leaf of the gate_sm
then we know that there are no implications.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>