ranges: fix a bug in rl_invert()
The first line which returns early for void types is, basically an
admission that I'm not sure what's going on here.  I don't think it
actually changes behavior.

The bug is the next part.  If we were supposed to add type_max to inverted
range then it wasn't doing that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>