implied: handle implications for if (({whatever; foo;}) < 0) {
Before we didn't handle the implications of expression statements like:

	if (({whatever; foo;}) < 0) {

We could handle it if you did:

	ret = ({whatever; foo;});
	if (ret < 0) {

But if the expression statement was part of the comparison, then we didn't
separate it out.  Now we do.

The one complication is that "foo" from this example is out of scope from
the comparison so we have already removed it from the slist.  I solved this
by delaying the removal a bit.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>