extra: fix handling of "if (foo & 0x3210) {"
If we have a condition like:

	if (foo & 0x3210) {

Then we know that on the true side that foo is at least 0x10.  So if we
start out knowing that foo is 0xfff then we know that foo is in the
0x10-0xff0 range.  The original code was just wrong.  What we need to
do is remove the range zero to the lowest bit minus one.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>