move freeing to AFTER_FUNC_HOOK
In olden times it was important to add everything to check_list.h in the
correct order so that you didn't use something that had been freed.  Now
we have the AFTER_FUNC_HOOK where it should be fine to free everything I
think?  I was trying to debug a use after free and just got sick of this
so I changed everything to use AFTER_FUNC_HOOK.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>