locking: add more impossible handling
The truth is that I have forgotten what this is about exactly or if it
solves any issues.  But Oleg is complaining about unspecified issues with
impossible conditions in check_locking.c and I have code which fixes
unspecified issues that I have not yet committed so maybe I can commit it
and make everyone happy.

I'm in the process of totally re-writing this check so hopefully that will
be ready soon.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>