extra: handle if (IS_ERR(foo[bar])) { ...
We could handle "if (IS_ERR(foo)) { ..." but not "if (IS_ERR(foo[bar])) {"
because the latter is a chunk and not a var sym.  This patch is more
limited than I would like because it doesn't handle PARAM_FILTER.

Also I modified set_extra_expr_nomod() to handle chunks.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>