function_hooks: integrate call_implies hooks with the db
I was trying to debug something which I think turned out to be caused by
something else.  But really, it's been a long standing issue that I have
been worrying about is that call_implies and the db weren't working
together.

How this works is that we basically fake another row in the db.  It assumes
that the db and the call_implies sort of agree on what the function
returns.

We have to keep the old code around for people who haven't built the cross
function db.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>