function_hooks: check PARAM_LIMIT implications all the time
Before then what was happening was that we were using the implications for
condition calls like "if (IS_ERR(p)) {".  But then there was a function
that did this:

void my_unlock(struct whatever *p)
{
	if (p->fast_path)
		return;
	spin_unlock(&p->lock);
}

Going into the function, we new that if ->fast_path was set then we were
not holding the lock but if it wasn't set then we were.  So we need to split
the implications apart to process this function correctly.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>