extra: handle fake assignments faster
Before we would do a search of the DB but that doesn't add anything so
for fake assignments only set the state if we have useful information.

From reading this patch, you might think that for other fake assignments
then we set it to the whole_rl() but actually smatch extra is already
supposed to search for the original state and if there is no original state
we don't bother setting the state to the whole range because that just
wastes memory and doesn't add information.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>