extra: deferences in fake assignments don't count
Sometimes we do a dereference when we're faking an assignment, but it
doesn't mean the pointer is non-NULL.  If we merge a pointer p where p
is NULL on one side and p->foo is set on the other side then we want to
preserve the value of p->foo.  So we have a fake dereference but it doesn't
mean anything.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>