extra: fix bug handling untracked pointers
Say we have some code like "unknown_function(p);" where unknown_function()
is not in the database.  The want to say that we don't know what p->foo is
after the end of the function call.  But unfortunately we were also
resetting the value of "p" which is a bug.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>