extra: export get_other_name_sym()
This is kind of a handy function when you are building the DB.  This is
already built into the smatch_param_*.c files because they use smatch
extra modification hooks.  But for check_locking sometimes people do
lock(param->foo->bar->lock) followed by an unlock(bar->lock);.  It looks
different but it's the same variable.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>