ptrlist: use after free in last_ptr_list()
This change is similar to 2e7dd34d11cb ('ptrlist: reading deleted items in
NEXT_PTR_LIST()').  If we use DELETE_CURRENT_PTR() then we can end up with
a list->nr that is zero meaning that we have to go back another list->prev
to find what we want.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>