function_hooks: fix locking false positives when you have no db
When I integrated the db and the call_implies hooks, I only tested it with
a database.  Fortunately, the validation tests don't have a DB so I caught
this bug before pushing it.

The issue is that we were always saying that the database handled it when,
in fact, there was no DB.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>