db/kernel.return_fixes: add scnprintf()
The scnprintf() doesn't return negative numbers.  It returns a positive
number less than size.  (Let's assume the size is not zero to make it
easier.  If the size is zero we deserve the problems we face).

Left to it's own devices Smatch thinks scnprintf() returns a few different
possible values so that's why there are three entries for this function.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>