db/kernel.return_fixes: hack around down_interruptible()
The problem here is that if you haven't rebuilt your database a few times
then smatch doesn't know what down_interruptible() returns and it breaks it
up into zero or anything.  The problem is that check_locking says anything
is both locked and unlocked, and zero is unlocked.  It gets confused and
prints a double unlock warning.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>