function_hooks: fix db and ranged_functions integration more
The problem here again is that I was testing with a pretty complete DB so
I didn't see this bug initially.  You have a bunch of cases where a
function returns negatives on error and 0 on success but for Smatch to see
that you need to rebuilt the DB several times.  The first time you build it
it just says s32min-s32max.  That can be both true and false so the ranged
function hooks are both run, overwriting each other so only the last state
is used.

If the ranged functions don't fit cleanly into true or false then pull them
out and handle them separately.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>