function_hooks: fix a type issue (affects ranged function hooks)
The issue here is that say you have:

	ret = some_function();

Imagine that "ret" is a u32 and some_function() returns int (negative
on error and zero on success).  In the original code, we stored the return
value as u32.  This was fine before but now we integrated with the ranged
function hooks we need to save it as int and cast it to u32 when we set
ret.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>