db: split on boolean parameter
This is quite a hack, but you know I'm not above that.  What my observation
is, is that we have a bunch of functions which do:

void whatever(int a, int b, int read_or_write) {

And depending on what read_or_write does, it behaves totally differently.

This code, does not change very much, because it tries to split on bool
as a last resort.  So most of the time it only affects void functions.
Also it only finds one bool param, it prefers the right most params.

It makes the DB slightly larger.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>