implied: fix a bug where we imply the wrong thing
This is complicated to explain without a diagram, maybe.  We use the
sm_in_keep_leafs() to see if we should preserve the whole sm state in
filter_stack.  But we should also use it in filter_pools() to see if we
should preserve a chunk of the sm_state.

Anyway, the result of this bug is that we don't preserve parts of the state
so say maybe that a value has to be zero but maybe it could be zero or any
positive number.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>