slist, implied: preserve the entire cur_stree in the pool for fake_strees
For fake strees, which are used in conditions and functions etc, we are
only interested in the states which are changed from when we first did
a push fake stree.  That's fine.

The problem is that the implied code, wants to see what the other states
were in the pool and that's not recorded.  So this patch creates a new
stree with that information.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>