db: mark some true false paths impossible based on db info
I've sort of re-ordered the changes from how I wrote them so the test case
I was using is not affected by this change yet because I haven't changed
smatch_db.c.  But the logic is still sound.

The situation is, say we come to a condition then we do get_implied_rl()
and we see if the returned rl is possibly true/false and mark the states
accordingly.  That's fine.  The problem arises because get_implied_rl()
looks up the returned values in the database and merges them together to
get the implied rl.  That's sounds fine, but sometimes those values are
only returned on impossible paths.

It would be better to fix get_implied_rl() to remove the impossible return
values but that seems complicated and probably slow.  So instead of that,
we get the same information from the function_hooks instead.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>