db: handle comparisons that we know are true better
Say you have a comparison:

	return foo->aaa == foo->bbb;

Say we know that those are not equal (the same thing applies for equal
as well).  It means that in call_return_state_hooks() then
get_implied_value() returns true, so it's not treated as a boolean
return but as a vanilla return of zero.

The problem is that we have merged the != and == states together and the
possible and impossible paths.  We want to separate them out so that we
can record in the database that foo->aaa != foo->bbb.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>