states: nullify_path() should do free the fake_cur_stree
This is a very core change so it's surprising how little difference it
makes in my testing.  Imagine you have this example code:

        /* w/a for a sporadic read returning 0 by waiting for the GT
         * thread to wake up.
         */
        if (({
                int ret, timeout = 500 * 1000;
                u64 base;

                base = local_clock();
                for (;;) {
                        u64 now = local_clock();
                        if ((__raw_i915_read32(dev_priv, GEN6_GT_THREAD_STATUS_REG) &
                            GEN6_GT_THREAD_STATUS_CORE_MASK) == 0) {
                                ret = 0;
                                break;
                        }
                        if (now - base >= timeout) {
                                ret = -ETIMEDOUT;
                                break;
                        }
                }
                ret;
            })) {
                DRM_ERROR("GT thread status wait timed out\n");
        }

The problem is in the original code, it complained that ret was not
initialized.  What's happening is that at the break statements we use the
cur_stree.  The cur_stree is the same as the fake_stree except with the
older states added as well.  Then we hit the break statement, and we
nullify the path and use the false conditions.  The issue is that now
first of all we don't really need the fake_cur stree because presumably
those states have been used.  Also now the cur_stree and fake_stree are
out of sync.

Solve this problem by freeing the fake_stree and starting a new one.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>