comparison: load/use return comparisons in the cross function db
This loads the db and also selects it and loads it back as well.

The thing about this is we're only considering the final state of the
parameter, not the initial state.  So if we set $0->end = $0->start then
we still return == even though they may have started out as not equal.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>